+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.0.0 (2022-09-19)
+
+### Features
+
+- add loading spinner and file upload control ([6190e7b](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/6190e7be675a98c1a74b89d43f494d42bdf5f598))
+- add onDelete interceptor ([710f937](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/710f93710bb67c42ab6b2fe5ba2075cef1fe1483))
+- add spinner ([d8ca826](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/d8ca82660fad7941e4da524ce0a4232faf2f97d6))
+- create upload compenent and stories ([b1569df](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/b1569df1388c520a1a6c6e6bff66eb39481828d5))
+- expose buttonProps to customize the upload button ([3702de1](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/3702de1035b0179952a94c168760e1bdeced8422))
+- setup storybook ([4e6a718](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/4e6a7180e3820cfa84c638dc8a5730a47512907d))
+
+### Bug Fixes
+
+- lint issues ([9c6b26b](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/9c6b26b030b71b2e346f84c6897a85c5e8cf2526))
+
+### 1.0.0 (2022-09-19)
+
+### Features
+
+- add loading spinner and file upload control ([6190e7b](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/6190e7be675a98c1a74b89d43f494d42bdf5f598))
+- add onDelete interceptor ([710f937](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/710f93710bb67c42ab6b2fe5ba2075cef1fe1483))
+- add spinner ([d8ca826](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/d8ca82660fad7941e4da524ce0a4232faf2f97d6))
+- create upload compenent and stories ([b1569df](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/b1569df1388c520a1a6c6e6bff66eb39481828d5))
+- expose buttonProps to customize the upload button ([3702de1](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/3702de1035b0179952a94c168760e1bdeced8422))
+- setup storybook ([4e6a718](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/4e6a7180e3820cfa84c638dc8a5730a47512907d))
+
+### Bug Fixes
+
+- lint issues ([9c6b26b](https://github.com/klinikum-evb/react-material-file-upload-spinner/commit/9c6b26b030b71b2e346f84c6897a85c5e8cf2526))
+
 ### [0.0.4](https://github.com/iamchathu/react-material-file-upload/compare/v0.0.3...v0.0.4) (2021-10-18)
 
 ### Features